requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements_common.in requirements_prod.in
#
cffi==1.16.0
    # via mozjpeg-lossless-optimization
cmake==3.27.4.1
    # via triton
filelock==3.12.4
    # via
    #   torch
    #   triton
jinja2==3.1.2
    # via torch
kornia==0.7.0
    # via -r requirements_common.in
lit==16.0.6
    # via triton
markupsafe==2.1.3
    # via jinja2
mozjpeg-lossless-optimization==1.1.3
    # via -r requirements_common.in
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
numpy==1.25.1
    # via
    #   -r requirements_common.in
    #   pandas
    #   scipy
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.1
    # via kornia
pandas==2.0.3
    # via -r requirements_common.in
pillow==9.3.0
    # via -r requirements_common.in
py4j==0.10.9.7
    # via pyspark
pycparser==2.21
    # via cffi
pyspark==3.4.1
    # via -r requirements_common.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
scipy==1.11.2
    # via -r requirements_common.in
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
torch==2.0.1
    # via
    #   -r requirements_prod.in
    #   kornia
    #   triton
triton==2.0.0
    # via torch
typing-extensions==4.7.1
    # via torch
tzdata==2023.3
    # via pandas
wheel==0.41.2
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11

# The following packages are considered to be unsafe in a requirements file:
# setuptools
